+++ src/Trips/PageTrips.tsx
@@ -12,16 +12,16 @@ import {
 import { DateTime } from 'luxon';
 import { useEffect, useState } from 'react';
 import { Link, type RouteComponentProps } from 'wouter';
-import { useCurrentUser } from '../../Auth/hooks';
-import { UserAvatarMenu } from '../../Auth/UserAvatarMenu';
-import { useBoundStore, useDeepBoundStore } from '../../data/store';
-import type { DbUser } from '../../data/types';
-import { DocTitle } from '../../Nav/DocTitle';
-import { Navbar } from '../../Nav/Navbar';
-import { RouteTrip } from '../../Routes/routes';
-import { TripGroup, type TripGroupType } from '../TripGroup';
-import { TripNewDialog } from '../TripNewDialog';
-import { formatTimestampToReadableDate } from '../time';
+import { useCurrentUser } from '../Auth/hooks';
+import { UserAvatarMenu } from '../Auth/UserAvatarMenu';
+import { useBoundStore, useDeepBoundStore } from '../data/store';
+import type { DbUser } from '../data/types';
+import { DocTitle } from '../Nav/DocTitle';
+import { Navbar } from '../Nav/Navbar';
+import { RouteTrip } from '../Routes/routes';
+import { TripGroup, type TripGroupType } from '../Trip/TripGroup';
+import { TripNewDialog } from '../Trip/TripNewDialog';
+import { formatTimestampToReadableDate } from '../Trip/time';
 import { useTripsGrouped } from './hooks';
 import s from './PageTrips.module.css';
 import type { TripsSliceTrip } from './store';
