+++ src/data/store.ts
@@ -2,8 +2,9 @@ import { create } from 'zustand';
 import { createUserSlice, type UserSlice } from '../Auth/store';
 import { createDialogSlice, type DialogSlice } from '../Dialog/hooks';
 import { createToastSlice, type ToastSlice } from '../Toast/hooks';
-import { createTripSlice, type TripSlice } from '../Trip/store';
-import { createTripsSlice, type TripsSlice } from '../Trip/Trips/store';
+import { createTripSlice } from '../Trip/store/store';
+import type { TripSlice } from '../Trip/store/types';
+import { createTripsSlice, type TripsSlice } from '../Trips/store';
 import { useDeepEqual } from './useDeepEqual';
 
 export type BoundStoreType = ToastSlice &
